### 2020-08-30 #2612 @whedon
[PRE REVIEW]: Psifr: Analysis and visualization of free recall data
**Submitting author:** @mortonne (<a href="http://orcid.org/0000-0002-2631-2710">Neal Morton</a>)
**Repository:** <a href="https://github.com/mortonne/psifr" target ="_blank">https://github.com/mortonne/psifr</a>
**Version:** v0.4.2
