### 2020-08-31 #2612 @mortonne
The article proof looks right.
