### 2020-09-10 #2612 @jeremymanning
@cMadan I’m super time-limited at the moment.  I could probably start taking a look sometime in October or November, so feel free to reach out then if you’re still having trouble finding reviewers.  But if you’re able to find someone who can turn this around faster I suggest going with them.  You might reach out to @andrewheusser and/or @paxtonfitzpatrick, who were both developers of our lab’s (related) [quail](cdl-quail.readthedocs.io) toolbox, also published in JOSS.
