### 2020-09-10 #2612 @samhforbes
It's a little outside of my area @cMadan, but it looks interesting and I'm happy to have a look if you're happy to have me.
