### 2020-09-14 671da41
docs: add badge for JOSS review to README
