### 2020-09-15 mortonne/psifr#1 @samhforbes
Analysis documentation
Hi @mortonne I'm really impressed with psifr so far, and really enjoying playing around with the various functions.
This is a (really) minor point
The documentation shows really clearly how to use the functions, which is great. I'm wondering if there are a few places where you could aid the user in adding a sentence or two on what the results mean. So as an example, in the function documentation for cat_crp (at the end of recall order), you've said that the aim is to test whether they recall items from the same category, and in the code chunk output a dataframe, and then the mean and sem. Although it's fairly clear what's coming out of this analysis, a sentence or two at the end that said something like: "From this we can see that the probability of ....." etc could really aid users who are less experienced with these analyses. Likewise for plots, a sentence saying what can be gleaned from the plots would be really welcome.
