### 2020-09-15 mortonne/psifr#2 @samhforbes
JOSS paper references
Hi @mortonne , the paper is also really well done. Minor issue here, please include DOIs for Murdock 1962 and Kahana 1996.

Another thing (and I want to point this is *entirely* up to you, is since some of the visualisations are a real strength of this package, you may wish to include one such plot in the paper as it gives the reader a taste of what the package can do, but as I say, that's entirely up to you.
