### 2020-09-15 #2669 @samhforbes
Hi @mortonne - thanks for the opportunity to look at this package. Installation worked smoothly, and while I have no free recall data of my own to test it on, the sample data provided works well (thanks for that). I thought the paper was very clear on the differences between this package and other existing packages, so that was also great. A minor thought (purely optional) is that you could consider puting one of your custom plots in the paper, since it is one of the strengths of this package.
I'm enjoying it so far!

I have opened a couple of issues for you to look at:
https://github.com/mortonne/psifr/issues/1
https://github.com/mortonne/psifr/issues/2
