### 2020-09-17 mortonne/psifr#5 @paxtonfitzpatrick
Add merge_free_recall() to "Scoring data" docs page
re: openjournals/joss-reviews#2669

The ["Scoring list recall" section of the docs](https://psifr.readthedocs.io/en/latest/guide/score.html#scoring-list-recall) provides a quick walkthrough of splitting a dataset into study and recall groups, and using the `merge_lists()` function to match presented and recalled items:
```
In [5]: study = data.query('trial_type == "study"').copy()

In [6]: recall = data.query('trial_type == "recall"').copy()

In [7]: merged = fr.merge_lists(study, recall)
```

which appears to be exactly what the `merge_free_recall()` function does:
https://github.com/mortonne/psifr/blob/671da414cbbb32615312da839a2f186e5cdfaf42/src/psifr/fr.py#L156-L166

It looks like `merge_free_recall()` was added a convenience function after that page of the docs was written, so it'd be great to update that section of the docs to at least note the function's existence.  Alternatively, the section could simply be reframed as an explanation of what `merge_free_recall()` does.
