### 2020-09-17 mortonne/psifr#4 @paxtonfitzpatrick
it'd also probably be worth pinning minimum versions of other packages whose APIs have had non-trivial changes lately (`pandas>=1.0.0`, `seaborn>=0.9.1`, etc.)
