### 2020-09-17 mortonne/psifr#6 @paxtonfitzpatrick
Tests for plotting functions
re: openjournals/joss-reviews#2669

Everything looks pretty well covered by the tests, with the exception of the 5 plotting functions.  Would it be possible to add some really simple tests for those? E.g., even something as quick as
```
def test_plot_raster(self):
    recall = fr.spc(self.data)
    g = fr.plot_spc(recall)
    plt.close()
    assert isinstance(g, sns.FacetGrid)
```
in [test_spc.py](https://github.com/mortonne/psifr/blob/671da414cbbb32615312da839a2f186e5cdfaf42/tests/test_spc.py) could help catch issues like #4.

You can get Travis CI to cooperate with plotting functions by adding the [xvfb](https://www.x.org/releases/X11R7.6/doc/man/man1/Xvfb.1.xhtml) service to `.travis.yml`
```
services:
  - xvfb
```
