### 2020-09-18 mortonne/psifr#7 @paxtonfitzpatrick
using filter_data() on columns with dtype=float
re: openjournals/joss-reviews#2669

The [`filter_data`](https://github.com/mortonne/psifr/blob/671da414cbbb32615312da839a2f186e5cdfaf42/src/psifr/fr.py#L26) function uses [`_match_value`](https://github.com/mortonne/psifr/blob/671da414cbbb32615312da839a2f186e5cdfaf42/src/psifr/fr.py#L17) to filter merged or un-merged data by a value or iterable of values in various attribute columns.  `_match_value` decides whether to use `pandas.Series.isin` or regular array broadcasting to create the mask by checking whether passed value is specifically an `int`.

In a merged dataset (i.e., after calling `fr.merge_free_recall(unmerged_data)`), missed items are assigned a `NaN` value in the `output` column and list intrusions are assigned a `NaN` in the input column, making those columns `dtype`s actually `float`s, rather than `int`s.  Filtering the merged data by a single `int` value in those columns (i.e., `fr.filter_data(merged_data, input=1)`) still works because `int(1)` and `float(1)` are considered equal, but filtering with a single `float` fails because the value gets passed to `pandas.Series.isin`, which takes a list-like object.

I think it'd be useful to handle float values in the filtering for a couple reasons:
1. users are likely to look at the DataFrame, see that the column values are clearly `float`s, and pass a `float` to `filter_data`
2. this would also allow passing `np.nan` as the single value in order to look specifically at either misses or intrusions

The simplest way to accomplish this would probably be to update `_match_values()` to either:
```
def _match_values(series, values):
    """Get matches for a data column."""
    if isinstance(values, (int, float, np.integer, np.floating)):
        include = series == values
    else:
        include = series.isin(values)
    return include
```
or
```
def _match_values(series, values):
    """Get matches for a data column."""
    if not hasattr(values, '__iter__') or isinstance(values, str):
        values = [values]
    include = series.isin(values)
    return include
```
