### 2020-09-18 mortonne/psifr#8 @paxtonfitzpatrick
Two last (minor) suggestions
re: openjournals/joss-reviews#2669

@mortonne thanks for putting together a great toolkit!  Here are a couple final (very minor) things I noticed and felt would be improvements.  Neither of them are essential to an initial release and I think `psifr` is worthy of publication in JOSS with or without these changes, so please feel free to treat this like a normal issue and take/leave whatever you'd like!

Thanks again!

## 1. add module imports to` __init__.py`
It'd be helpful to be able to use `psifr` functions by referencing them under the name of the package itself, rather than individually importing its submodules or functions, since both modules' names and most of the functions' names are fairly non-specific.  For example, `from psifr import fr` makes it easy to accidentally overwrite `fr` with a variable later, and `from psifr.fr import filter_data` can make it hard to tell where `filter_data` came from.  Being able to call `psifr.fr.filter_data()` would be helpful for a user IMO.

However, the package namespace is constructed a little oddly right now, such that what's available in the namespace changes based on whether or not you've imported other things already. If you just `import psifr`, none of `psifr`'s functionality is actually accessible:
```
In [1]: import psifr

In [2]: dir(psifr)
Out[2]:
['__builtins__',
 '__cached__',
 '__doc__',
 '__file__',
 '__loader__',
 '__name__',
 '__package__',
 '__path__',
 '__spec__']
```
meaning that trying to use a `psifr` function as `psifr.module.function` fails:
```
In [1]: import psifr

In [2]: psifr.fr.sample_data('Morton2013').head()
<ipython-input-2-5e1d412ed05e> in <module>
----> 1 psifr.fr.sample_data('Morton2013').head()

AttributeError: module 'psifr' has no attribute 'fr'
```

Somewhat confusingly, however,  since both the `fr` and `transitions` modules import each other, this actually works if you import either one of the modules first:
```
In [1]: from psifr import transitions

In [2]: import psifr

In [3]: dir(psifr)
Out[3]:
['__builtins__',
 '__cached__',
 '__doc__',
 '__file__',
 '__loader__',
 '__name__',
 '__package__',
 '__path__',
 '__spec__',
 'fr',
 'transitions']

In [4]: psifr.fr.sample_data('Morton2013').head()
Out[4]:
   subject  list  position trial_type      item  item_number  session list_type category  response  response_time list_category
0        1     1         1      study     TOWEL          743        1      pure      obj       3.0          1.517           obj
1        1     1         2      study     LADLE          631        1      pure      obj       3.0          1.404           obj
2        1     1         3      study   THERMOS          735        1      pure      obj       3.0          0.911           obj
3        1     1         4      study      LEGO          637        1      pure      obj       3.0          0.883           obj
4        1     1         5      study  BACKPACK          521        1      pure      obj       3.0          0.819           obj
```

This could be easily fixed and made consistent by adding
```
import psifr.fr
import psifr.transitions
```
to `__init__.py`.

Ultimately, all this would change is allowing you to access functions as `psifr.module.function()` without importing an individual module/function first, but it would make this an explicit feature rather than a side-effect of how the package is set up.

## 2. change `**kwargs` to `**facet_kws` in `plot_spc`
All of the plotting functions forward their keyword arguments to `sns.FacetGrid`, but this is the only one whose header calls them `kwargs` rather than `facet_kws`.  For consistency (and to help clarify what these kwargs are passed to, since it isn't explicitly mentioned in the docstring), it would probably be worth changing this to match the others.
