### 2020-09-18 #2669 @paxtonfitzpatrick
Hey @mortonne, thanks for putting together a great toolkit!  I've finished looking everything over, and apart from a few fixes and tweaks in the issues I linked above, everything ran smoothly, looked pretty, and was well documented 🙂
I agree with @samhforbes that a plot or two (particularly one of the raster plots, in my opinion) would make a great addition to the paper if you wanted!  But of course that's entirely up to you, and I think `psifr` is worthy of publication in JOSS with or without it.

I mentioned this in the issue itself, but I just wanted to note here as well that I don't think addressing the two items in mortonne/psifr#8 is necessary for an initial release, and both are likely outside the scope of a JOSS review (though I’ll defer to @cMadan on that since this is my first time not on the receiving end).  I simply enjoyed playing around with `psifr` and could certainly see myself using it, so I thought I'd post a couple quick suggestions I felt would be improvements.

Thanks again, and I'm signed off pending issues 1-7!
