### 2020-09-18 mortonne/psifr#8 @cMadan
I agree with @paxtonfitzpatrick that these are useful improvements, but that the acceptance of the JOSS submission does not require them to be implemented first.
