### 2020-09-18 #2669 @cMadan
@paxtonfitzpatrick, thank you for the thorough review! I've commented in https://github.com/mortonne/psifr/issues/8, but to briefly re-cap here--I agree this is a useful improvement but is not necessary for JOSS acceptance.
