+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{Murdock:1962,
     author = {Murdock, Bennet B},
     title = {{The serial position effect of free recall}},
+    doi = {10.1037/h0045106},
     pages = {482 488},
     number = {5},
     volume = {64},
@@ -11,6 +12,7 @@
 @article{Kahana:1996,
     author = {Kahana, Michael Jacob},
     title = {{Associative retrieval processes in free recall}},
+    doi = {10.3758/bf03197276},
     pages = {103 109},
     number = {1},
     volume = {24},
