### 2020-09-19 a6a3db5
feat: support for brief legends in raster plots
