### 2020-09-19 5792343
feat: support for setting legend verbosity in raster plots
