### 2020-09-19 cf30c5a
docs: add legend to raster plot
