### 2020-09-19 5ce31a5
docs: adjust subject spc to make it more compact
