### 2020-09-19 96e2e9e
docs: add sample figures to paper.md
