### 2020-09-19 3b5feb7
docs: use svg for embedded images
