### 2020-09-19 401a492
docs: use pdf for embedded images
