### 2020-09-19 da90510
docs: fix figure references
