### 2020-09-19 c8e2db7
docs: explain how intrusions are visualized
