### 2020-09-21 mortonne/psifr#2 @mortonne
Thanks, this is a great suggestion. I've added plots of recall by serial position, both on average and for individual participants. I've also added an example of a raster plot, to highlight a type of plot I haven't seen done in similar packages.

I have also added DOIs for those references.
