### 2020-09-21 51c517e
docs: change plot file format from png to svg
