### 2020-09-21 58e8e97
docs: update installation instructions in README.md
