+++ README.md
@@ -26,17 +26,17 @@ See the [user guide](https://psifr.readthedocs.io/en/latest/guide/overview.html)
 
 ## Installation
 
-First get a copy of the code from GitHub:
+You can install the latest stable version of Psifr using pip:
 
 ```bash
-git clone git@github.com:mortonne/psifr.git
+pip install psifr
 ```
 
-Then install:
+You can also install the development version directly from the code
+repository on GitHub:
 
 ```bash
-cd psifr
-python setup.py install
+pip install git+git://github.com/mortonne/psifr
 ```
 
 ## Importing data
