### 2020-09-21 8a0ac36
build: use pip for installation
