+++ .travis.yml
@@ -3,7 +3,7 @@ python:
   - "3.8"
 install:
   - pip install -r requirements.txt
-  - python setup.py develop
+  - pip install .
 script:
   - pytest --cov-report term --cov=psifr
 after_success:
