### 2020-09-21 c441791
docs: use simpler method for scoring free recall data
