### 2020-09-21 7e94a2a
docs: simplify formatting of python code
