+++ docs/guide/recall.rst
@@ -20,13 +20,11 @@ Recall performance
 
 First, load some sample data and create a merged DataFrame:
 
-.. ipython::
+.. ipython:: python
 
-    In [1]: from psifr import fr
-
-    In [1]: df = fr.sample_data('Morton2013')
-
-    In [1]: data = fr.merge_free_recall(df)
+    from psifr import fr
+    df = fr.sample_data('Morton2013')
+    data = fr.merge_free_recall(df)
 
 Raster plot
 ~~~~~~~~~~~
@@ -37,12 +35,11 @@ colored by the serial position of the recalled item in the list. Items near
 the end of the list are shown in yellow, and items near the beginning of the
 list are shown in purple. Intrusions of items not on the list are shown in red.
 
-.. ipython::
+.. ipython:: python
 
-    In [1]: subj = fr.filter_data(data, 1)
-
+    subj = fr.filter_data(data, 1)
     @savefig raster_subject.svg
-    In [1]: g = fr.plot_raster(subj).add_legend()
+    g = fr.plot_raster(subj).add_legend()
 
 Serial position curve
 ~~~~~~~~~~~~~~~~~~~~~
@@ -50,20 +47,19 @@ Serial position curve
 We can calculate average recall for each serial position
 using :py:func:`~psifr.fr.spc` and plot using :py:func:`~psifr.fr.plot_spc`.
 
-.. ipython::
+.. ipython:: python
 
-    In [1]: recall = fr.spc(data)
-
+    recall = fr.spc(data)
     @savefig spc.svg
-    In [1]: g = fr.plot_spc(recall)
+    g = fr.plot_spc(recall)
 
 Using the same plotting function, we can plot the curve for each
 individual subject:
 
-.. ipython::
+.. ipython:: python
 
     @savefig spc_indiv.svg
-    In [1]: g = fr.plot_spc(recall, col='subject', col_wrap=5)
+    g = fr.plot_spc(recall, col='subject', col_wrap=5)
 
 Probability of Nth recall
 ~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -71,11 +67,10 @@ Probability of Nth recall
 We can also split up recalls, to test for example how likely participants
 were to initiate recall with the last item on the list.
 
-.. ipython::
+.. ipython:: python
 
-    In [1]: prob = fr.pnr(data)
-
-    In [1]: prob
+    prob = fr.pnr(data)
+    prob
 
 This gives us the probability of recall by output position (:code:`'output'`)
 and serial or input position (:code:`'input'`). This is a lot to look at all
@@ -84,9 +79,8 @@ We can plot the curves using :py:func:`~psifr.fr.plot_spc`, which takes an
 optional :code:`hue` input to specify a variable to use to split the data
 into curves of different colors.
 
-.. ipython::
+.. ipython:: python
 
-    In [1]: pfr = prob.query('output <= 3')
-
+    pfr = prob.query('output <= 3')
     @savefig pnr.svg
-    In [1]: g = fr.plot_spc(pfr, hue='output').add_legend()
+    g = fr.plot_spc(pfr, hue='output').add_legend()
