### 2020-09-21 345064d
docs: update information about custom columns
