### 2020-09-21 mortonne/psifr#5 @mortonne
Most users won't need to customize which events go into the study and recall bins, so I've updated the scoring page to use `merge_free_recall` instead. I've also added a reference to another page to better explain how to include custom columns.
