### 2020-09-21 f2f2f81
docs: add local installation instructions
