### 2020-09-21 mortonne/psifr#3 @mortonne
Thanks for pointing this out. I've had users run into problems that may have been related to the uninstallation problem you mentioned.
* Local installation is mostly relevant for developers, so I've moved that from `README.md` to `.github/CONTRIBUTING.md`. I've updated the instructions for local installation there.
* I've updated the installation instructions to match the documentation page instructions.
* The CI build configuration has been updated to use `pip`.

I'm leaving this issue open until I make a new release on PyPI, which will update the installation directions there.
