### 2020-09-21 4026678
build: specify matplotlib!=3.3.1 to avoid raster problem
Closes #4.
