+++ requirements.txt
@@ -1,6 +1,7 @@
 numpy
 scipy
 pandas
+matplotlib!=3.3.1
 seaborn
 pytest
 pytest-cov
