### 2020-09-21 76943c3
build: remove install_requires
Remove setup.py requirements so they are only set in requirements.txt.
