+++ setup.py
@@ -21,12 +21,6 @@ setuptools.setup(
     package_data={
         'psifr': ['data/*.csv']
     },
-    install_requires=[
-        'numpy',
-        'pandas',
-        'matplotlib!=3.3.1',
-        'seaborn',
-    ],
     classifiers=[
         'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
         'Programming Language :: Python :: 3.8',
