### 2020-09-21 84df3fc
build: add requirements for docs and testing
