### 2020-09-21 b8151db
build: remove testing packages from requirements.txt
