### 2020-09-21 c8891cb
build: pin minimal versions of pandas and seaborn
