### 2020-09-21 9ae0487
build: add back install_requires list to setup.py
