### 2020-09-21 028ec17
build: use extra requirements for test in travis build
