+++ .travis.yml
@@ -2,7 +2,7 @@ language: python
 python:
   - "3.8"
 install:
-  - pip install -r requirements.txt
+  - pip install -r requirements.txt .[test]
   - pip install .
 script:
   - pytest --cov-report term --cov=psifr
