### 2020-09-21 aaad7ea
build: switch to using requirements for rtd build
