+++ .readthedocs.yml
@@ -19,6 +19,9 @@ formats: all
 # Optionally set the version of Python and requirements required to build your docs
 python:
   version: 3.8
-
-conda:
-  environment: environment.yml
+  install:
+    - requirements: requirements.txt
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
