### 2020-09-21 44f86ff
build: remove unused environment.yml file
