+++ environment.yml
@@ -1,85 +1,0 @@
-name: psifr
-channels:
-  - defaults
-dependencies:
-  - alabaster=0.7.12
-  - asn1crypto=1.3.0
-  - babel=2.8.0
-  - backcall=0.1.0
-  - blas=1.0
-  - ca-certificates=2020.1.1
-  - certifi=2019.11.28
-  - cffi=1.14.0
-  - chardet=3.0.4
-  - cryptography=2.8
-  - cycler=0.10.0
-  - decorator=4.4.2
-  - docutils=0.16
-  - freetype=2.9.1
-  - idna=2.9
-  - imagesize=1.2.0
-  - intel-openmp=2019.4
-  - ipython=7.13.0
-  - ipython_genutils=0.2.0
-  - jedi=0.16.0
-  - jinja2=2.11.1
-  - kiwisolver=1.0.1
-  - libedit=3.1.20181209
-  - libffi=3.2.1
-  - libpng=1.6.37
-  - markupsafe=1.1.1
-  - matplotlib=3.1.3
-  - matplotlib-base=3.1.3
-  - mkl=2019.4
-  - mkl-service=2.3.0
-  - mkl_fft=1.0.15
-  - mkl_random=1.1.0
-  - ncurses=6.2
-  - numpy=1.18.1
-  - numpy-base=1.18.1
-  - openssl=1.1.1e
-  - packaging=20.3
-  - pandas=1.0.3
-  - parso=0.6.2
-  - pexpect=4.8.0
-  - pickleshare=0.7.5
-  - pip=20.0.2
-  - prompt-toolkit=3.0.4
-  - prompt_toolkit=3.0.4
-  - ptyprocess=0.6.0
-  - pycparser=2.20
-  - pygments=2.6.1
-  - pyopenssl=19.1.0
-  - pyparsing=2.4.6
-  - pysocks=1.7.1
-  - python=3.8.2
-  - python-dateutil=2.8.1
-  - pytz=2019.3
-  - readline=8.0
-  - requests=2.23.0
-  - scipy=1.4.1
-  - seaborn=0.10.0
-  - setuptools=46.1.3
-  - six=1.14.0
-  - snowballstemmer=2.0.0
-  - sphinx=2.4.4
-  - sphinxcontrib-applehelp=1.0.2
-  - sphinxcontrib-devhelp=1.0.2
-  - sphinxcontrib-htmlhelp=1.0.3
-  - sphinxcontrib-jsmath=1.0.1
-  - sphinxcontrib-qthelp=1.0.3
-  - sphinxcontrib-serializinghtml=1.1.4
-  - sqlite=3.31.1
-  - tk=8.6.8
-  - tornado=6.0.4
-  - traitlets=4.3.3
-  - urllib3=1.25.8
-  - wcwidth=0.1.9
-  - wheel=0.34.2
-  - xz=5.2.4
-  - zlib=1.2.11
-  - pip:
-    - pydata-sphinx-theme==0.3.0
-    - git+https://github.com/mortonne/psifr
-prefix: //anaconda3/envs/psifr
-
