### 2020-09-21 e9c42ee
docs: fix formatting of sample data
