### 2020-09-21 a795ece
test: set up separate raw and data fixtures
