+++ tests/test_fr.py
@@ -7,7 +7,7 @@ from psifr import fr
 
 
 @pytest.fixture()
-def frame():
+def raw():
     raw = pd.DataFrame(
         {'subject': [1, 1, 1, 1, 1, 1,
                      1, 1, 1, 1, 1, 1],
@@ -27,10 +27,14 @@ def frame():
                         3, 4, 5, 5, 3, 5],
          'task': [1, 2, 1, 2, 1, np.nan,
                   1, 2, 1, 1, 1, 1]})
-    study = raw.query('trial_type == "study"').copy()
-    recall = raw.query('trial_type == "recall"').copy()
-    data = fr.merge_lists(study, recall, study_keys=['task'],
-                          list_keys=['item_index'])
+    return raw
+
+
+@pytest.fixture()
+def data(raw):
+    data = fr.merge_free_recall(
+        raw, study_keys=['task'], list_keys=['item_index']
+    )
     return data
 
 
@@ -45,29 +49,29 @@ def distances():
     return mat
 
 
-def test_split_lists(frame):
-    study = fr.split_lists(frame, 'study', ['item', 'input', 'task'])
+def test_split_lists(data):
+    study = fr.split_lists(data, 'study', ['item', 'input', 'task'])
     np.testing.assert_allclose(study['input'][1], np.array([1., 2., 3.]))
-    recall = fr.split_lists(frame, 'recall', ['input'], ['recalls'])
+    recall = fr.split_lists(data, 'recall', ['input'], ['recalls'])
     np.testing.assert_allclose(recall['recalls'][0], np.array([2., 3., np.nan]))
 
 
-def test_pnr(frame):
-    stat = fr.pnr(frame)
+def test_pnr(data):
+    stat = fr.pnr(data)
     expected = np.array([0, 0.5, 0.5, 0.5, 0, 1, np.nan, np.nan, np.nan])
     observed = stat['prob'].to_numpy()
     np.testing.assert_array_equal(expected, observed)
 
 
-def test_lag_rank(frame):
-    stat = fr.lag_rank(frame)
+def test_lag_rank(data):
+    stat = fr.lag_rank(data)
     expected = np.array([0.25])
     observed = stat['rank'].to_numpy()
     np.testing.assert_array_equal(expected, observed)
 
 
-def test_distance_rank(frame, distances):
-    stat = fr.distance_rank(frame, 'item_index', distances)
+def test_distance_rank(data, distances):
+    stat = fr.distance_rank(data, 'item_index', distances)
     expected = np.array([0.25])
     observed = stat['rank'].to_numpy()
     np.testing.assert_array_equal(expected, observed)
