### 2020-09-21 4814b93
test: filtering of raw and merged data
