### 2020-09-21 0147c9c
test: basic plotting checks
