### 2020-09-21 f3827e2
build: add xvfb service for plot testing
Refs: #6
