+++ .travis.yml
@@ -8,3 +8,5 @@ script:
   - pytest --cov-report term --cov=psifr
 after_success:
   - codecov
+services:
+  - xvfb
