### 2020-09-21 eafd0de
test: vertical raster plot
Refs: #6
