### 2020-09-21 c34ddd7
test: min_samples argument in plot_distance_crp
Refs: #6
