### 2020-09-21 efc5805
refactor: rename kwargs to facet_kws for clarity
Refs: #8
