+++ src/psifr/__init__.py
@@ -1,0 +1,2 @@
+import psifr.fr
+import psifr.transitions
