### 2020-09-21 6c6afa4
refactor: fix for filtering with individual floats
Refs: #7
