### 2020-09-21 7cb66f3
test: filtering with individual floats
Refs: #7
