### 2020-09-21 mortonne/psifr#7 @mortonne
I've implemented your second suggested implementation and added a unit test to confirm that it works.
