### 2020-09-21 28d0c07
docs: add more explanation of analysis interpretation
Refs: #1
