### 2020-09-21 mortonne/psifr#1 @mortonne
Hi @samhforbes thanks very much for taking the time to look at the project! I've added additional explanation to the documentation to explain how measures and plots should be interpreted.
