### 2020-09-21 mortonne/psifr#3 @mortonne
Installation instructions have been updated on GitHub and PyPI.
