### 2020-09-21 mortonne/psifr#8 @mortonne
Thanks for the thoughtful comments! I've implemented your last two suggestions.
