### 2020-09-23 227f0ac
docs: add quickstart section to README.md
