### 2020-09-24 #2669 @samhforbes
Hi @mortonne - great work with these issues - I'm really happy with how you have updated the docs, I think it will be good for the reader. The tests for plotting functions mentioned by @paxtonfitzpatrick are great, and the high test coverage is now really impressive. From my end I'm certainly satisfied. Well done, it's a great package, and I will be recommending it.
