### 2020-09-24 mortonne/psifr#6 @paxtonfitzpatrick
ah yeah, sorry, I should've noted `seaborn==0.10.1` added a workaround on their end for the underlying matplotlib issue.  It happened for me using `matplotlib==3.3.1` and `seaborn==0.9.1`.  Leaving `matplotlib!=3.3.1` pinned seems like the right way to go, since the pinned seaborn version is prior to their fix being implemented.

Thanks for handling this, and the tests look great to me!
