### 2020-09-24 #2669 @paxtonfitzpatrick
@mortonne thanks very much for addressing these issues—all your fixes look great to me!  I'm more than happy with the changes you've made, and am looking forward both to using `psifr` myself and seeing how the package grows in the future.
