### 2020-10-14 #2669 @mortonne
* I've created a new tagged release: [v0.4.4](https://github.com/mortonne/psifr/releases/tag/v0.4.4).
* That tagged release has been archived in [Zenodo](https://zenodo.org/record/4086188); the archive has the correct metadata.
* Zenodo DOI: 10.5281/zenodo.4086188
