### 2020-10-14 #2669 @cMadan
@mortonne, can you proof your paper.md one last time. In particular, you should revise this line:
``The source code for `Psifr` has  been archived to Zenodo with the linked DOI: TBD.``
