### 2020-10-14 67bd7fd
docs: update paper with zenodo DOI
