### 2020-10-14 #2669 @mortonne
I've added the Zenodo DOI and proofed the paper. The archive and JOSS DOI links give 404 errors, but I'm guessing that's normal at this stage. Everything else looks good.
